feat(editLink): Adapted relevant edit link to target core repo
